requirements.txt
asgiref==3.6.0
autopep8==2.0.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.5
django-admin-list-filter-dropdown==1.0.3
django-admin-rangefilter==0.9.0
django-ckeditor==6.5.1
django-grappelli==3.0.4
django-import-export==3.0.2
django-js-asset==2.0.0
et-xmlfile==1.1.0
Faker==16.6.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.4.0
pycodestyle==2.10.0
python-dateutil==2.8.2
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
tablib==3.3.0
tomli==2.0.1
xlrd==2.0.1
xlwt==1.3.0
